mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.52.0",
    "mcp>=1.12.0",
    "python-dotenv>=1.1.1",
]
